emitElf: Fix possible sign-extension errors
When you have an expression type "<uint64_t> = <uint16_t> * <uint16_t>",
it sadly takes a detour into signed values.  The operands will be
promoted to int, then multiplied, then sign-extended to finally fill
uint64_t.  If the multiplication result is greater than 0x7fffffff,
overflowing int, then the top 32 bits of the uint64_t will all be 1s.
(That's if the compiler doesn't blow up the int overflow as undefined.)

The easiest solution is to forcefully promote the operands to the target
size and sign before multiplying.

Reported by Coverity as SIGN_EXTENSION.

+++ symtabAPI/src/emitElf.C
@@ -1172,7 +1172,7 @@ void emitElf::fixPhdrs(unsigned &extraAlignSize)
   // libelf from overwriting the program headers data when outputing
   // sections.  Fill in the new section's data with what we just wrote.
   Elf_Data *data = elf_newdata(phdrs_scn);
-  size_t total_size = newEhdr->e_phnum * newEhdr->e_phentsize;
+  size_t total_size = (size_t)newEhdr->e_phnum * (size_t)newEhdr->e_phentsize;
   data->d_buf = malloc(total_size);
   memcpy(data->d_buf, phdr_data, total_size);
   data->d_size = total_size;
